LayoutTests/imported/w3c/web-platform-tests/svg/geometry/parsing/rx-valid-expected.txt

CONSOLE MESSAGE: line 96: TypeError: null is not an object (evaluating 'document.body.appendChild')
#PID UNRESPONSIVE - WebKitTestRunner (pid 53737)
FAIL: Timed out waiting for notifyDone to be called

#EOF
#EOF
